// src/pages/protect/types/protectPortal.types.ts
export interface TokenSet {
	accessToken: string;
	tokenType: string;
	expiresIn: number;
	scope: string;
	idToken?: string;
	refreshToken?: string;
}